+++ cmd/vega/node/node_pre.go
@@ -39,6 +39,7 @@ import (
 	"code.vegaprotocol.io/vega/snapshot"
 	"code.vegaprotocol.io/vega/spam"
 	"code.vegaprotocol.io/vega/staking"
+	"code.vegaprotocol.io/vega/statevar"
 	"code.vegaprotocol.io/vega/stats"
 	"code.vegaprotocol.io/vega/subscribers"
 	"code.vegaprotocol.io/vega/types"
@@ -274,6 +275,7 @@ func (l *NodeCommand) startBlockchain(ctx context.Context, commander *nodewallet
 		l.stakingAccounts,
 		l.rewards,
 		l.snapshot,
+		l.statevar,
 		l.Version,
 	)
 
@@ -339,11 +341,6 @@ func (l *NodeCommand) preRun(_ []string) (err error) {
 	l.timeService.NotifyOnTick(l.oracle.UpdateCurrentTime)
 	l.oracleAdaptors = oracleAdaptors.New()
 
-	// instantiate the execution engine
-	l.executionEngine = execution.NewEngine(
-		l.Log, l.conf.Execution, l.timeService, l.collateral, l.oracle, l.broker,
-	)
-
 	// we cannot pass the Chain dependency here (that's set by the blockchain)
 	commander, err := nodewallets.NewCommander(l.conf.NodeWallet, l.Log, nil, l.nodeWallets.Vega, l.stats)
 	if err != nil {
@@ -361,6 +358,13 @@ func (l *NodeCommand) preRun(_ []string) (err error) {
 		l.Log, l.conf.Staking, l.broker, l.timeService, l.witness, l.ethClient, l.netParams,
 	)
 	l.epochService = epochtime.NewService(l.Log, l.conf.Epoch, l.timeService, l.broker)
+
+	l.statevar = statevar.New(l.Log, l.conf.StateVar, l.broker, l.topology, commander, l.timeService)
+
+	// instantiate the execution engine
+	l.executionEngine = execution.NewEngine(
+		l.Log, l.conf.Execution, l.timeService, l.collateral, l.oracle, l.broker,
+	)
 
 	if l.conf.Blockchain.ChainProvider == blockchain.ProviderNullChain {
 		// Use staking-loop to pretend a dummy builtin asssets deposited with the faucet was staked
@@ -411,7 +415,6 @@ func (l *NodeCommand) preRun(_ []string) (err error) {
 	if err != nil {
 		panic(err)
 	}
-
 	// notify delegation, rewards, and accounting on changes in the validator pub key
 	l.topology.NotifyOnKeyChange(l.delegation.ValidatorKeyChanged, l.stakingAccounts.ValidatorKeyChanged, l.rewards.ValidatorKeyChanged, l.governance.ValidatorKeyChanged)
 
@@ -599,6 +602,14 @@ func (l *NodeCommand) setupNetParameters() error {
 		netparams.WatchParam{
 			Param:   netparams.SnapshotIntervalLength,
 			Watcher: l.snapshot.OnSnapshotIntervalUpdate,
+		},
+		netparams.WatchParam{
+			Param:   netparams.ValidatorsVoteRequired,
+			Watcher: l.statevar.OnDefaultValidatorsVoteRequiredUpdate,
+		},
+		netparams.WatchParam{
+			Param:   netparams.FloatingPointUpdatesDuration,
+			Watcher: l.statevar.OnFloatingPointUpdatesDurationUpdate,
 		},
 	)
 }